Explicit return type

+++ src/dsprot_main.c
@@ -59,7 +59,7 @@ u32 DSProt_DetectFlashcart(void* callback) {
 	
 	ENCRYPTION_END(KEY_DSPROT_MAIN_1);
 	
-	return result;
+	return (u32)result;
 }
 
 
@@ -99,7 +99,7 @@ u32 DSProt_DetectNotFlashcart(void* callback) {
 	
 	ENCRYPTION_END(KEY_DSPROT_MAIN_2);
 	
-	return result;
+	return (u32)result;
 }
 
 
@@ -139,7 +139,7 @@ u32 DSProt_DetectEmulator(void* callback) {
 	
 	ENCRYPTION_END(KEY_DSPROT_MAIN_3);
 	
-	return result;
+	return (u32)result;
 }
 
 
@@ -179,7 +179,7 @@ u32 DSProt_DetectNotEmulator(void* callback) {
 	
 	ENCRYPTION_END(KEY_DSPROT_MAIN_4);
 	
-	return result;
+	return (u32)result;
 }
 
 
@@ -219,7 +219,7 @@ u32 DSProt_DetectDummy(void* callback) {
 	
 	ENCRYPTION_END(KEY_DSPROT_MAIN_5);
 	
-	return result;
+	return (u32)result;
 }
 
 
@@ -259,5 +259,5 @@ u32 DSProt_DetectNotDummy(void* callback) {
 	
 	ENCRYPTION_END(KEY_DSPROT_MAIN_6);
 	
-	return result;
+	return (u32)result;
 }
